Fix wishlist num
