+++ src/main/java/com/finalproject/everrent_be/domain/member/service/MemberService.java
@@ -62,9 +62,7 @@ public class MemberService {
     @Transactional
     public Member getMemberfromContext() {
         String userId = SecurityContextHolder.getContext().getAuthentication().getName();
-        System.out.println(userId);
         Optional<Member> member = memberRepository.findById(Long.valueOf(userId));  //Long.valueOf(userId)
-        System.out.println(member.get());
         return member.get();
     }
